docs: removing customization file contradiction
Customising via the configuration file is clearly possible when using yaml files since you show an example for this. This PR updates the doc to remove the warning that states this is not possible. Given that yaml was the only missing file type in the original warning it seems safe to remove it completely.