fix(pre-commit): Confine hook to commit-msg stage
The Commitizen pre-commit hook runs cz check, which only makes sense to
do on a commit message, not at other Git hook stages. By default the
hook is currently run on staged files pre-commit, which is nonsensical
unless the files in the repository themselves contain Git commit
messages. The hook is an implausible candidate even for the
prepare-commit-msg stage since this stage is intended for hooks that
modify the commit message rather than check it. Specifying the hook
stage centrally makes it less error-prone to use and saves one line of
configuration for users of the pre-commit hook.