fix(bump): Support regexes containing colons
The version_files feature uses a colon to delimit filenames from a regex
to use to find the version. Make version_files slightly more robust by
splitting on the first colon rather than all colons, thereby permitting
the regex to contain a colon. Maintain the pre-existing assumption that
the filename doesn't contain a colon, as this is rare and would likely
cause many other problems.