fix(bump): Search for version number line by line
Avoid the complexities of multiline regex matching and inconsistencies
with the case where no regex is specified. Simplify the implementation
by eliminating the need to track an offset into the file. Ensure that
the version number will be found even if it is to the left of the
portion of the line matching the regex.