ci(pre-commit): Remove version_files lookahead
Commitizen runs bump on itself to, among other things, manage the
version it uses of its own pre-commit hooks. When searching
.pre-commit-config.yaml for the version number of Commitizen, we search
for "Commitizen," which is in a comment to the right of the version
number. Remove the lookahead assertion for "Commitizen," and instead
simply consume it when matching now that version_files regexes don't
need to exclusively match to the left of the version number.