test(bump): handle Windows path separator
Windows uses `\` for path separation, causing python to see the `\U` in
`C:\Users` as a Unicode string. Replace `\\` with `/` in file path
strings.

Also, the colon in drive paths on Windows causes the `partition` logic
in `bump.py` and `commands/bump.py` to fail. Perform the logic on the
path without the drive portion.