fix(changelog): allow rev range lookups without a tag format
The current default setting for `tag_format` is `None`. This is not a
problem for the `bump` command, since the `normalize_tag` function
defaults to `$version` when no `tag_format` is passed. However it is a
problem for the `changelog` command, which seems to explicitly demand a
`tag_format` in order to run a rev-range lookup. This creates issues
like https://github.com/commitizen-tools/commitizen/issues/622.

Either a sane default needs to be set for `tag_format` or the
restriction in `changelog` has to be uplifted. In this commit the latter
has been chosen. A test is also implemented to check that `changelog`
will always compute a rev range with the default tag format.

Fixes https://github.com/commitizen-tools/commitizen/issues/622