feat: Make `cz bump` work on shallow clones
This allows reliably setting up your CI to only fetch the latest changes.
E.g. Gitlab CI uses shallow clones with a depth of 20 by default:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/77576

Only `tag_exists()` had to be modified for this. `cz changelog` is not
ready yet, more functions would have to include a `git fetch` for that.

I was worried about slowing down error reporting in non-CI workflows ("git
fetch" calls can sometimes take a while if your connection is bad). To
minimize the risk for this, now only fetching the tag for shallow
clones and always printing a warning when this happens.

I was hit by this issue while trying out Commitizen in Gitlab CI on a
monorepo, where some packages haven't seen a release for >50 commits.