feat(bump): support prereleases with start offset
In special cases, it may be necessary for prereleases not to start with a 0, because the individual characters are encoded in bytes. Here, a 0 in the byte is interpreted as a release version, consistent with the location of bugfixes. Therefore, this commit introduces a generic option to change the start of the prerelease to accommodate such circumstances.

EXAMPLES

Before:
0.3.0 -> PATCH beta-> 0.3.1b0

Now (with offset 0 == default)
0.3.0 -> PATCH beta-> 0.3.1b0

Now (with offset 1)
0.3.0 -> PATCH beta-> 0.3.1b1