feat(providers): add support for `Cargo.lock`
If `Cargo.lock` file is present in the same path as the `Cargo.toml`,
the version of the package will also be updated in the `Cargo.lock`
file.

The implementation is similar to the one used for `UvProvider`, which
extends the `CargoProvider` to handle lock files. The change does not
break any existing functionality, if the `Cargo.lock` file is not
present, since it checks for the existence of the file before
attempting a lock file update.